AMP-5080
git-svn-id: file:///svn/trunk@8454 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/action/Login.java
@@ -260,7 +260,7 @@ public class Login extends Action {
 					appSettings.setAppSettingsId(ampAppSettings
 							.getAmpAppSettingsId());
 					appSettings.setDefRecsPerPage(ampAppSettings
-							.getDefaultRecordsPerPage().intValue());
+							.getDefaultRecordsPerPage());
 					appSettings.setCurrencyId(ampAppSettings.getCurrency()
 							.getAmpCurrencyId());
 					appSettings.setFisCalId(ampAppSettings.getFiscalCalendar()
